+++ packages/checkout-extension/src/Extension.tsx
@@ -1,10 +1,10 @@
 import { type ExtensionRegion } from '@bigcommerce/checkout-sdk';
 import React, { type ReactNode, useEffect } from 'react';
 
+import { useExtensions } from '@bigcommerce/checkout/contexts';
 import { useCheckout } from '@bigcommerce/checkout/payment-integration-api';
 
 import { extensionRegionToContainerMap } from './ExtensionRegionContainer';
-import { useExtensions } from './useExtensions';
 
 interface ExtensionProps {
     region: ExtensionRegion;