+++ packages/contexts/src/extension/index.ts
@@ -1,0 +1,5 @@
+export { ExtensionProvider } from './ExtensionProvider';
+export { ExtensionContextProps } from './ExtensionContext';
+export { withExtension } from './withExtension';
+export { useExtensions } from './useExtensions';
+export { ExtensionAction, ExtensionServiceInterface, ExtensionActionType } from './ExtensionType';