+++ packages/core/src/app/shipping/MultiShippingForm.test.tsx
@@ -7,8 +7,8 @@ import {
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import { ExtensionProvider } from '@bigcommerce/checkout/checkout-extension';
-import { type ErrorLogger } from '@bigcommerce/checkout/error-handling-utils';
+import { ExtensionService } from '@bigcommerce/checkout/checkout-extension';
+import { ExtensionProvider ,type ExtensionServiceInterface } from '@bigcommerce/checkout/contexts';
 import {
     createLocaleContext,
     LocaleContext,
@@ -33,25 +33,24 @@ import { getShippingAddress } from './shipping-addresses.mock';
 
 describe('MultiShippingForm Component', () => {
     let checkoutService: CheckoutService;
+    let extensionService: ExtensionServiceInterface;
     let checkoutState: CheckoutSelectors;
     let defaultProps: MultiShippingFormProps;
     let localeContext: LocaleContextType;
-    let errorLogger: ErrorLogger;
 
     beforeEach(() => {
         localeContext = createLocaleContext(getStoreConfig());
         checkoutService = createCheckoutService();
+        extensionService = new ExtensionService(checkoutService, createErrorLogger());
         checkoutState = checkoutService.getState();
 
         defaultProps = {
+            cartHasChanged: false,
             customerMessage: 'x',
-            countriesWithAutocomplete: [],
             isLoading: false,
             onUnhandledError: jest.fn(),
-            onSubmit: jest.fn(),
+            onSubmit: jest.fn()
         };
-
-        errorLogger = createErrorLogger();
 
         jest.spyOn(checkoutState.data, 'getBillingAddressFields').mockReturnValue(
             getAddressFormFields(),
@@ -90,7 +89,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -122,7 +121,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -170,7 +169,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -187,7 +186,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -246,7 +245,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -372,7 +371,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -458,7 +457,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -503,7 +502,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -554,7 +553,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -621,7 +620,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -684,7 +683,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>
@@ -741,7 +740,7 @@ describe('MultiShippingForm Component', () => {
         render(
             <CheckoutProvider checkoutService={checkoutService}>
                 <LocaleContext.Provider value={localeContext}>
-                    <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger} >
+                    <ExtensionProvider extensionService={extensionService}>
                         <MultiShippingForm {...defaultProps} />
                     </ExtensionProvider>
                 </LocaleContext.Provider>