+++ packages/core/src/app/shipping/stripeUPE/StripeShipping.test.tsx
@@ -4,8 +4,8 @@ import {
 } from '@bigcommerce/checkout-sdk';
 import React from 'react';
 
-import { ExtensionProvider } from '@bigcommerce/checkout/checkout-extension';
-import { type AnalyticsEvents, AnalyticsProviderMock } from '@bigcommerce/checkout/contexts';
+import { ExtensionService } from '@bigcommerce/checkout/checkout-extension';
+import { type AnalyticsEvents, AnalyticsProviderMock, ExtensionProvider } from '@bigcommerce/checkout/contexts';
 import { LocaleProvider } from '@bigcommerce/checkout/locale';
 import { CheckoutProvider } from '@bigcommerce/checkout/payment-integration-api';
 import { render } from '@bigcommerce/checkout/test-utils';
@@ -16,13 +16,13 @@ import CheckoutStepType from '../../checkout/CheckoutStepType';
 import ConsoleErrorLogger from '../../common/error/ConsoleErrorLogger';
 import { getStoreConfig } from '../../config/config.mock';
 import { getCustomer } from '../../customer/customers.mock';
-import { getShippingAddress } from '../shipping-addresses.mock';
 
 import StripeShipping, { type StripeShippingProps } from './StripeShipping';
 
 describe('Stripe Shipping Component', () => {
     const checkoutService = createCheckoutService();
     const errorLogger = new ConsoleErrorLogger();
+    const extensionService = new ExtensionService(checkoutService, errorLogger);
     const initialize = jest.fn();
     let checkoutState: CheckoutSelectors;
     let analyticsTracker: Partial<AnalyticsEvents>;
@@ -30,6 +30,7 @@ describe('Stripe Shipping Component', () => {
    checkoutService.initializeShipping = initialize;
 
   const defaultProps: StripeShippingProps = {
+        isInitializing: false,
         step: {
             isActive: true,
             isBusy: false,
@@ -67,7 +68,7 @@ describe('Stripe Shipping Component', () => {
         <CheckoutProvider checkoutService={checkoutService}>
           <LocaleProvider checkoutService={checkoutService}>
             <AnalyticsProviderMock analyticsTracker={analyticsTracker}>
-              <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger}>
+                <ExtensionProvider extensionService={extensionService}>
                 <StripeShipping {...defaultProps} />
               </ExtensionProvider>
             </AnalyticsProviderMock>